move test helpers to a separate file
